+++ src/kudu/util/async_logger.h
@@ -86,7 +86,7 @@ class AsyncLogger : public google::base::Logger {
   void Write(bool force_flush,
              time_t timestamp,
              const char* message,
-             int message_len) override;
+             size_t message_len) override;
 
   // Flush any buffered messages.
   void Flush() override;
